requirements.txt
# DeepFusion Requirements
# Object Detection Only - Jetson AGX Orin Deployment

# Core Dependencies
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
opencv-python>=4.5.0
PyYAML>=6.0
tqdm>=4.62.0
matplotlib>=3.5.0

# Optional: For ONNX export
onnx>=1.12.0
onnxruntime>=1.12.0

# Optional: For TensorRT (Jetson)
# tensorrt>=8.5.0  # Install via JetPack

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Visualization
pillow>=9.0.0
# shapely>=1.8.0  # For accurate IoU computation

# Jupyter (for experimentation)
# jupyter>=1.0.0
# ipywidgets>=7.6.0
